feat: migrate code execution from Firebase Cloud Tasks to BullMQ
Replace onTaskDispatched handlers with a BullMQ-based architecture where
Firebase Functions only handle enqueueing and scoring, while Docker execution
runs on self-hosted workers outside Firebase.

- Add submissions.ts: onSubmissionCreated enqueues jobs to BullMQ `decathlon` queue
- Add executions.ts: onExecutionCreated reads results from Firestore `executions`
  collection and applies game-specific scoring/ranking logic
- Remove esolang.ts (split into submissions.ts and executions.ts)
- Add Execution and ExecutionResultItem types to schema.d.ts
- Add bullmq, ioredis dependencies; remove axios
- Workers write to executions/{submissionId} after Docker execution,
  decoupling execution from Firebase billing and removing task timeout constraints
