+++ functions/package.json
@@ -17,9 +17,10 @@
   "main": "lib/index.js",
   "dependencies": {
     "@slack/web-api": "^7.15.1",
-    "axios": "^1.15.2",
+    "bullmq": "^5.0.0",
     "firebase-admin": "^13.8.0",
     "firebase-functions": "^7.2.5",
+    "ioredis": "^5.0.0",
     "mdiff": "^1.4.2",
     "remeda": "^2.33.7"
   },
